content should show before sidebar in HTML
It is good (SEO) practice to show the content before showing the sidebar in the HTML markup. In Foundation you can do so by using push/pull. Push is used to push a div to the right and Pull is used to pull a div (back) to the left.

I have made these changes to your side-menu.php, most likely you will need to make them too in other templates.